+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # [Unreleased]
 
+* Changed: replaced `log::error` with `eprintln`
+* Changed: improved the error messages
+
 # [0.5.1] - 2022-10-14
 
 * Charged: Replaced `time` dep with `chrono`